+++ lib/widgets/circular_indicator.dart
@@ -4,7 +4,6 @@
     createTime:2021-08-26 19:15
  */
 
-// import 'dart:io' show Platform;
 import 'package:universal_io/io.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';